+++ tests/testthat.R
@@ -1,5 +1,0 @@
-library(dbplyr)
-library(sergeant)
-library(testthat)
-
-test_check("sergeant")